Add method AddMargin(). It makes margin for bbox of a tile
